+++ inspirehep/records/ext.py
@@ -19,4 +19,4 @@ class InspireRecords(object):
     def init_app(self, app):
         """Flask application initialization."""
         app.extensions["inspirehep-records"] = self
-        from inspirehep.records.api import receivers  # noqa: F401
+        from inspirehep.records import receivers  # noqa: F401
